+++ packages/core/navigation-next/src/components/connected/LayoutManagerWithViewController/LayoutManagerWithViewController.js
@@ -1,209 +1,47 @@
 // @flow
 
-import React, {
-  Component,
-  Fragment,
-  type ElementConfig,
-  type Node,
-} from 'react';
-import { NavigationAnalyticsContext } from '@atlaskit/analytics-namespaced-context';
+import React from 'react';
 
-import { ViewControllerSubscriber } from '../../../view-controller';
-import ItemsRenderer from '../../../renderer';
-import { withNavigationUIController } from '../../../ui-controller';
-import LayoutManager from '../../presentational/LayoutManager';
-import type {
-  LayoutManagerWithViewControllerProps,
-  LayoutManagerWithViewControllerState,
-} from './types';
+import type { LayoutManagerWithViewControllerProps } from './types';
+import ViewRenderer from '../../../renderer';
 import SkeletonContainerView from '../../presentational/SkeletonContainerView';
-import type { ActiveView } from '../../../view-controller/types';
-import LayerInitialised from '../../presentational/LayerInitialised';
+import AsyncLayoutManagerWithViewController from '../AsyncLayoutManagerWithViewController';
 /* NOTE: experimental props use an underscore */
 /* eslint-disable camelcase */
 
-class LayoutManagerWithViewControllerBase extends Component<
-  LayoutManagerWithViewControllerProps,
-  LayoutManagerWithViewControllerState,
-> {
-  static defaultProps = {
-    experimental_flyoutOnHover: false,
-  };
+const LayoutManagerWithViewController = ({
+  children,
+  firstSkeletonToRender,
+  customComponents,
+  experimental_flyoutOnHover,
+  globalNavigation,
+  onExpandStart,
+  onExpandEnd,
+  onCollapseStart,
+  onCollapseEnd,
+  getRefs,
+}: LayoutManagerWithViewControllerProps) => {
+  return (
+    <AsyncLayoutManagerWithViewController
+      onExpandStart={onExpandStart}
+      onExpandEnd={onExpandEnd}
+      onCollapseStart={onCollapseStart}
+      onCollapseEnd={onCollapseEnd}
+      getRefs={getRefs}
+      customComponents={customComponents}
+      experimental_flyoutOnHover={!!experimental_flyoutOnHover}
+      globalNavigation={globalNavigation}
+      containerSkeleton={() =>
+        firstSkeletonToRender ? (
+          <SkeletonContainerView type={firstSkeletonToRender} />
+        ) : null
+      }
+      viewRenderer={ViewRenderer}
+      firstSkeletonToRender={firstSkeletonToRender}
+    >
+      {children}
+    </AsyncLayoutManagerWithViewController>
+  );
+};
 
-  state = {
-    hasInitialised: false,
-    outgoingView: null,
-  };
-
-  constructor(props: LayoutManagerWithViewControllerProps) {
-    super(props);
-    this.renderContainerNavigation.displayName = 'ContainerNavigationRenderer';
-    this.renderProductNavigation.displayName = 'ProductNavigationRenderer';
-  }
-
-  componentDidUpdate(prevProps: LayoutManagerWithViewControllerProps) {
-    const { view } = this.props;
-    const { view: prevView } = prevProps;
-
-    if (!view || !prevView) {
-      return;
-    }
-
-    // If we're moving from a product to a container view or vice versa we cache
-    // the previous view so that we can still render it during the transition.
-    if (view.type !== prevView.type) {
-      // It's totally fine to setState in componentDidUpdate as long as it's
-      // wrapped in a condition:
-      // https://reactjs.org/docs/react-component.html#componentdidupdate
-      // eslint-disable-next-line
-      this.setState({ outgoingView: prevView });
-    }
-  }
-
-  onInitialised = () => {
-    this.setState({
-      hasInitialised: true,
-    });
-  };
-
-  renderSkeleton = () => {
-    return <SkeletonContainerView type={this.props.firstSkeletonToRender} />;
-  };
-
-  renderContainerNavigation = () => {
-    const { firstSkeletonToRender, view } = this.props;
-    const { outgoingView } = this.state;
-
-    if (view && view.type === 'container') {
-      return this.renderView(view);
-    }
-
-    if (outgoingView && outgoingView.type === 'container') {
-      return this.renderView(outgoingView);
-    }
-
-    if (
-      !view &&
-      firstSkeletonToRender === 'container' &&
-      !this.state.hasInitialised
-    ) {
-      return this.renderSkeleton();
-    }
-
-    return firstSkeletonToRender !== 'container' ? null : this.renderSkeleton();
-  };
-
-  renderGlobalNavigation = () => {
-    const { globalNavigation: GlobalNavigation, view } = this.props;
-    const { hasInitialised } = this.state;
-
-    /* We are embedding the LayerInitialised analytics component within global navigation so that
-     * the event it fires can access the analytics context within LayerManager. The component
-     * cannot be rendered directly within LayerManager since it needs access to view data which
-     * only exists in LayoutManagerWithViewController. */
-    return (
-      <Fragment>
-        <GlobalNavigation />
-        <LayerInitialised
-          activeView={view}
-          initialised={hasInitialised}
-          onInitialised={this.onInitialised}
-        />
-      </Fragment>
-    );
-  };
-
-  renderProductNavigation = () => {
-    const { view } = this.props;
-    const { outgoingView } = this.state;
-
-    if (view && view.type === 'product') {
-      return this.renderView(view);
-    }
-
-    // If we're transitioning from a product view to a container view still
-    // render the outgoing product view.
-    if (
-      view &&
-      view.type === 'container' &&
-      outgoingView &&
-      outgoingView.type === 'product'
-    ) {
-      return this.renderView(outgoingView);
-    }
-
-    return this.renderSkeleton();
-  };
-
-  renderView(view: ActiveView): Node {
-    const { customComponents } = this.props;
-    return (
-      <ItemsRenderer customComponents={customComponents} items={view.data} />
-    );
-  }
-
-  render() {
-    const {
-      children,
-      experimental_flyoutOnHover,
-      firstSkeletonToRender,
-      onExpandStart,
-      onExpandEnd,
-      onCollapseStart,
-      onCollapseEnd,
-      getRefs,
-      view,
-    } = this.props;
-
-    const shouldRenderContainerLayer =
-      (view && view.type === 'container') ||
-      (!view &&
-        firstSkeletonToRender === 'container' &&
-        !this.state.hasInitialised);
-
-    return (
-      <NavigationAnalyticsContext
-        data={{
-          attributes: {
-            navigationLayer: view && view.type,
-            view: view && view.id,
-            ...(view && view.analyticsAttributes),
-          },
-        }}
-      >
-        <LayoutManager
-          globalNavigation={this.renderGlobalNavigation}
-          containerNavigation={
-            shouldRenderContainerLayer ? this.renderContainerNavigation : null
-          }
-          experimental_flyoutOnHover={experimental_flyoutOnHover}
-          productNavigation={this.renderProductNavigation}
-          onExpandStart={onExpandStart}
-          onExpandEnd={onExpandEnd}
-          onCollapseStart={onCollapseStart}
-          onCollapseEnd={onCollapseEnd}
-          getRefs={getRefs}
-        >
-          {children}
-        </LayoutManager>
-      </NavigationAnalyticsContext>
-    );
-  }
-}
-
-const LayoutManagerWithView = (
-  props: $Diff<
-    ElementConfig<typeof LayoutManagerWithViewControllerBase>,
-    { view: * },
-  >,
-) => (
-  <ViewControllerSubscriber>
-    {({ state: { activeView } }) => {
-      return (
-        <LayoutManagerWithViewControllerBase view={activeView} {...props} />
-      );
-    }}
-  </ViewControllerSubscriber>
-);
-
-export default withNavigationUIController(LayoutManagerWithView);
+export default LayoutManagerWithViewController;