+++ packages/core/navigation-next/src/ui-controller/index.js
@@ -5,4 +5,3 @@ export { default as UIControllerSubscriber } from './UIControllerSubscriber';
 export {
   default as withNavigationUIController,
 } from './withNavigationUIController';
-export type { UIControllerInterface } from './types';