+++ packages/editor/editor-core/src/__tests__/unit/utils/document.ts
@@ -1,5 +1,4 @@
 import { name } from '../../../../package.json';
-import { Node } from 'prosemirror-model';
 import {
   doc,
   p,
@@ -20,7 +19,6 @@ import {
   isEmptyNode,
   isEmptyParagraph,
   isEmptyDocument,
-  preprocessDoc,
   processRawValue,
   hasVisibleContent,
 } from '../../../utils/document';
@@ -121,30 +119,6 @@ describe(name, () => {
 
       it('should return false if node has hr', () => {
         expect(isEmptyDocument(doc(p(), hr())(schema))).toBe(false);
-      });
-    });
-
-    describe('preprocessDoc', () => {
-      it('should return true if node is empty', () => {
-        const editorContent = doc(
-          p('some text'),
-          decisionList({})(decisionItem({})()),
-        )(schema);
-        const processedContent = preprocessDoc(schema, editorContent);
-        expect(processedContent).not.toBe(undefined);
-        expect((processedContent as Node)!.content!.childCount).toEqual(1);
-        expect(
-          (processedContent as Node)!.content!.firstChild!.type.name,
-        ).toEqual('paragraph');
-      });
-
-      it('should return new document', () => {
-        const editorContent = doc(
-          p('some text'),
-          decisionList({})(decisionItem({})()),
-        )(schema);
-        const processedContent = preprocessDoc(schema, editorContent);
-        expect(processedContent).not.toEqual(editorContent);
       });
     });
   });