+++ packages/editor/editor-wikimarkup-transformer/src/parser/builder/list-builder.ts
@@ -154,10 +154,8 @@ export class ListBuilder {
         continue;
       }
 
-      if (
-        pmNode.type.name === 'paragraph' &&
-        pmNode.textContent.trim() === ''
-      ) {
+      /* Skip Empty spaces after rule */
+      if (this.isParagraphEmptyTextNode(pmNode)) {
         continue;
       }
 
@@ -179,6 +177,23 @@ export class ListBuilder {
 
     return output;
   };
+
+  /* Check if all paragraph's children nodes are text and empty */
+  private isParagraphEmptyTextNode(node: PMNode): boolean {
+    if (node.type.name !== 'paragraph' || !node.childCount) {
+      return false;
+    }
+    for (let i = 0; i < node.childCount; i++) {
+      const n = node.content.child(i);
+      if (n.type.name !== 'text') {
+        // Paragraph contains non-text node, so not empty
+        return false;
+      } else if (n.textContent.trim() !== '') {
+        return false;
+      }
+    }
+    return true;
+  }
 
   private createListItem(content: PMNode[], schema: Schema): PMNode {
     if (