+++ packages/search/quick-search/src/components/QuickSearch.tsx
@@ -101,6 +101,8 @@ export type Props = {
   value?: string;
   /** Corresponds to the `resultId` of the selected result */
   selectedResultId?: SelectedResultId;
+  /** Optional way of being notified when the selected result changes due to keyboard nav */
+  onSelectedResultIdChanged?: (id: SelectedResultId) => void;
   // Internal: injected by withAnalytics(). Fire a private analytics event
   firePrivateAnalyticsEvent?: FireAnalyticsEvent;
   /** React component to be used for rendering links */
@@ -251,6 +253,9 @@ export class QuickSearch extends React.Component<Props, State> {
     });
     if (selectedResultId) {
       this.fireKeyboardControlEvent(selectedResultId);
+    }
+    if (this.props.onSelectedResultIdChanged) {
+      this.props.onSelectedResultIdChanged(selectedResultId);
     }
   };